subtotal and total methods

+++ coursespring/src/main/java/com/MatheusResio/coursespring/entities/Payment.java
@@ -20,7 +20,7 @@ public class Payment implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private Instant moment;
 
 	@JsonIgnore
@@ -31,18 +31,18 @@ public class Payment implements Serializable {
 	public Payment() {
 	}
 
-	public Payment(Long id, Instant moment, Order order) {
+	public Payment(Integer id, Instant moment, Order order) {
 		super();
 		this.id = id;
 		this.moment = moment;
 		this.order = order;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
